Clean up on failure
Updated version of my previous PR which was almost entirely reverted.
The bulk of the work is now done in a bash -e block which the cleanup code
is safely outside of, so the build will abort on errors, and be able to
report them more clearly, but junk files will still be removed afterwards.

I got rid of the "alias sudo=eval", as it wouldn't have worked anyway.
A better workaround for if the user doesn't have sudo would be nice.